refactor(SoupStrainer): split out standalone methods to trivial base
